Update DeepQNet.py
corrected errors in replay function
it used the target model to evaluate the current state, which should be done by the online model.
It did not use the batch correctly, as it itterated over the samples instead of passing them all through the model at the same time.
Increased the replay buffer size to 20,000